+++ components/api-server/src/config.js
@@ -10,7 +10,7 @@ module.exports = config;
 _.merge(config.schema, {
   singleCoreUrl: {
     format: String,
-    default: null,
+    default: '',
     doc: 'URL indicating the address to reach the server. When the platform operates as a single core, with register as a component. Setting this will deactivate some features and override some parameters.'
   },
   service: {