+++ components/api-server/src/methods/webhooks.js
@@ -53,23 +53,8 @@ module.exports = function produceWebhooksApiMethods(
 
   api.register('webhooks.get',
     commonFns.getParamsValidation(methodsSchema.get.params),
-    forbidSharedAccess,
     findAccessibleWebhooks,
   );
-
-  function forbidSharedAccess(context: MethodContext, params: mixed, result: Result, next: ApiCallback) {
-    const currentAccess: Access = context.access;
-
-    if (currentAccess == null) {
-      return next(new Error('AF: Access cannot be null at this point.'));
-    }
-    if (currentAccess.isShared()) {
-      return next(errors.forbidden(
-        'Shared Accesses cannot manipulate Webhooks. Please use an App or Personnal Access.'
-      ));
-    }
-    next();
-  }
 
   async function findAccessibleWebhooks(context: MethodContext, params: mixed, result: Result, next: ApiCallback) {
     const currentAccess = context.access;
@@ -88,7 +73,6 @@ module.exports = function produceWebhooksApiMethods(
 
   api.register('webhooks.getOne',
     commonFns.getParamsValidation(methodsSchema.get.params),
-    forbidSharedAccess,
     findWebhook,
   );
 
@@ -103,7 +87,7 @@ module.exports = function produceWebhooksApiMethods(
         return next(errors.unknownResource('webhook', params.id));
       }
       if (!isWebhookInScope(webhook, currentAccess)) {
-        return next(errors.forbidden('The webhook was not created by this app access.'));
+        return next(errors.forbidden('The webhook was not created by this access.'));
       }
 
       result.webhook = webhook.forApi();
@@ -117,9 +101,7 @@ module.exports = function produceWebhooksApiMethods(
 
   api.register('webhooks.create',
     commonFns.getParamsValidation(methodsSchema.create.params),
-    forbidSharedAccess,
     forbidPersonalAccess,
-    forbidCreateOnlyAccess,
     createWebhook,
     bootWebhook,
   );
@@ -133,20 +115,6 @@ module.exports = function produceWebhooksApiMethods(
     if (currentAccess.isPersonal()) {
       return next(errors.forbidden(
         'Personal Accesses cannot create Webhooks. Please use an App Access.'
-      ));
-    }
-    next();
-  }
-
-  function forbidCreateOnlyAccess(context: MethodContext, params: mixed, result: Result, next: ApiCallback) {
-    const currentAccess: Access = context.access;
-
-    if (currentAccess == null) {
-      return next(new Error('AF: Access cannot be null at this point.'));
-    }
-    if (currentAccess.hasCreateOnlyPermission()) {
-      return next(errors.forbidden(
-        'Accesses containing "create-only" permissions cannot create Webhooks. Please use another App Access.'
       ));
     }
     next();
@@ -190,7 +158,6 @@ module.exports = function produceWebhooksApiMethods(
 
   api.register('webhooks.update',
     commonFns.getParamsValidation(methodsSchema.update.params),
-    forbidSharedAccess,
     commonFns.catchForbiddenUpdate(webhookSchema('update'), false, logger),
     applyPrerequisitesForUpdate,
     updateWebhook,
@@ -246,7 +213,6 @@ module.exports = function produceWebhooksApiMethods(
 
   api.register('webhooks.delete',
     commonFns.getParamsValidation(methodsSchema.del.params),
-    forbidSharedAccess,
     deleteAccess,
     turnOffWebhook,
   );
@@ -296,7 +262,6 @@ module.exports = function produceWebhooksApiMethods(
 
   api.register('webhooks.test',
     commonFns.getParamsValidation(methodsSchema.test.params),
-    forbidSharedAccess,
     testWebhook,
   );
 
@@ -336,7 +301,7 @@ module.exports = function produceWebhooksApiMethods(
    */
   function isWebhookInScope(webhook: Webhook, access: Access): boolean {
     if (access.isPersonal()) return true;
-    return access.isApp() && access.id === webhook.accessId;
+    return access.id === webhook.accessId;
   }
   
 };